Borongan: treat max_reject as percentage

+++ database/migrations/2025_12_24_125203_create_borongan_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('kategori');
             $table->char('nama_item');
             $table->integer('satuan');
-            $table->integer('max_rej_subkon');
+            $table->double('max_rej_subkon');
 
             $table->boolean('status_aktif')->default(true);
